RGB-SPD files
RGB-SPD files uploaded (NOTE: there seems to be some issues in this version as the the values cannot be cross-checked, so it needs to be double-checked.) But the difference is not big.

There were two versions of calibration files (one measured with the square opening, the other one measured with a 10-degree circular opening), the values are slightly different.